requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.2.0
altair==5.0.1
antlr4-python3-runtime==4.9.3
asttokens==2.0.5
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
boto3==1.34.45
botocore==1.34.45
Bottleneck==1.3.7
Brotli==1.0.9
cachetools==4.2.2
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
comm==0.1.2
contourpy==1.2.0
cryptography==42.0.2
cycler==0.10.0
Cython==3.0.8
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.0
executing==0.8.3
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.0
fsspec==2024.2.0
fvcore==0.1.5.post20221221
gitdb==4.0.7
GitPython==3.1.37
gmpy2==2.1.2
google-auth==2.6.0
google-auth-oauthlib==0.5.2
groundingdino==0.1.0
grpcio==1.48.2
huggingface-hub==0.20.3
idna==2.10
imageio==2.33.1
importlib-metadata==6.0.0
iopath==0.1.10
ipykernel==6.28.0
ipython==8.20.0
ipywidgets==8.0.4
jedi==0.18.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lapx==0.5.5
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.6
net-tools==0.1.2
networkx==3.1
numexpr==2.8.7
numpy==1.26.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
packaging==23.1
pafy==0.5.5
pandas==2.1.4
parso==0.8.3
pexpect==4.8.0
Pillow==9.4.0
pip==23.3.1
platformdirs==3.10.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.7
pycparser==2.21
pydeck==0.8.0b4
Pygments==2.15.1
PyJWT==2.4.0
Pympler==0.9
pyOpenSSL==24.0.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.0
requests-toolbelt==1.0.0
rich==13.3.5
roboflow==1.1.19
rpds-py==0.10.6
rsa==4.7.2
s3transfer==0.10.0
safetensors==0.4.2
scikit-learn==1.3.0
scipy==1.11.4
seaborn==0.13.2
setuptools==68.2.2
six==1.16.0
smmap==4.0.0
stack-data==0.2.0
streamlit==1.31.1
supervision==0.18.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==2.2.0
timm==0.9.12
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.2.0
torchaudio==2.2.0
torchsummary==1.5.1
torchvision==0.17.0
tornado==6.3.3
tqdm==4.66.2
traitlets==5.7.1
transformers==4.38.1
triton==2.2.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==2.1
ultralytics==8.1.16
urllib3==2.0.7
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
Werkzeug==2.2.3
wget==3.2
wheel==0.41.2
widgetsnbextension==4.0.5
yacs==0.1.8
yapf==0.40.2
yarl==1.9.3
youtube-dl==2020.12.2
zipp==3.17.0
